backend/pyproject.toml
[project]
name = "project-grandson"
version = "0.1.0"
description = "Improved version of claude desktop backend using OpenInterpeter and vision"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyautogui>=0.9.54",
    "python-dotenv>=1.0.1",
    "aiohttp~=3.10.3",
    "loguru~=0.7.2",
    "Markdown~=3.7",
    "numpy~=1.26.4",
    "Pillow~=10.4.0",
    "protobuf~=4.25.4",
    "pydantic~=2.8.2",
    "pyloudnorm~=0.1.1",
    "resampy~=0.4.3",
    "pipecat-ai[anthropic,cartesia,daily,openai,silero]",
    "fastapi[standard]>=0.115.5",
]


[project.optional-dependencies]
dev = [
    "build~=1.2.1",
    "grpcio-tools~=1.62.2", 
    "pip-tools~=7.4.1",
    "pyright~=1.1.376",
    "pytest~=8.3.2",
    "ruff~=0.6.7",
    "setuptools~=72.2.0",
    "setuptools_scm~=8.1.0",
]

[tool.uv]
dev-dependencies = [
    "watchdog>=6.0.0",
]


[tool.uv.sources]
pipecat-ai = { path = "pipecat", editable = true }
